+++ CHANGELOG.md
@@ -1,6 +1,7 @@
 # Changelog
 
 ## v1 rework with go-chi
+- 2024-06-28 : fix stacked bars apex chart
 - 2024-06-27 : stats, new apex chart, stacked bars with possible group by category
 - 2024-06-23 : fix responsive apex chart
 - 2024-06-10 : added save routes